refactor(dispatch/act/refactor): extract plugin response handling into separate module
- Moved response handling and diff forwarding logic from mod.rs to a new response_handling.rs file
- Updated mod.rs to use the new response_handling module
- Added new refusal reason ProviderLacksCapability in resolution.rs for better provider capability checks

This change improves code organization and separation of concerns in the act refactor dispatch logic.
